fix: cast SUM() aggregate to BIGINT to fix type mismatch
PostgreSQL SUM() returns NUMERIC, not BIGINT, causing runtime error:
'error occurred while decoding column storage_bytes: mismatched types;
Rust type i64 (INT8) is not compatible with SQL type NUMERIC'

Changes:
- Cast SUM(file_size_bytes)::BIGINT in workspace_service_impl.rs
- Added PostgreSQL type gotchas section to migration safety guide
- Updated migration template with aggregate function warnings

Root Cause: PostgreSQL changes aggregate function return types to
prevent overflow (SUM(BIGINT) → NUMERIC, AVG(INT) → NUMERIC).

Impact: Fixes rebuild knowledge graph endpoint and workspace stats
queries. Ensures embedding type changes can rebuild without errors.

OODA-iteration: Type compatibility fix
Ref: #storage_bytes #rebuild_kg #type_safety

+++ edgequake/crates/edgequake-core/src/workspace_service_impl.rs
@@ -643,7 +643,7 @@ impl WorkspaceService for WorkspaceServiceImpl {
                 (SELECT COUNT(*) FROM entities WHERE workspace_id = $1) as entity_count,
                 (SELECT COUNT(*) FROM relationships WHERE workspace_id = $1) as relationship_count,
                 (SELECT COUNT(*) FROM chunks WHERE workspace_id = $1 AND embedding IS NOT NULL) as embedding_count,
-                (SELECT COALESCE(SUM(file_size_bytes), 0) FROM documents WHERE workspace_id = $1) as storage_bytes
+                (SELECT COALESCE(SUM(file_size_bytes), 0)::BIGINT FROM documents WHERE workspace_id = $1) as storage_bytes
             "#,
         )
         .bind(workspace_id)
